+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.1] - 2021-10-22
+
 ### Fixed
 - Various overlays that were depending on specific versions in nixpkgs 20.09.